Added LlmClient; refactoring
