+++ src/EntwineLLM/source.extension.vsixmanifest
@@ -1,26 +1,26 @@
 <?xml version="1.0" encoding="utf-8"?>
 <PackageManifest Version="2.0.0" xmlns="http://schemas.microsoft.com/developer/vsx-schema/2011" xmlns:d="http://schemas.microsoft.com/developer/vsx-schema-design/2011">
-    <Metadata>
-        <Identity Id="EntwineLlm.77601dfd-d66f-405b-b3c4-c2e05497a5ca" Version="1.10" Language="en-US" Publisher="Emiliano Musso" />
-        <DisplayName>EntwineLlm</DisplayName>
-        <Description xml:space="preserve">LLM coding assistant extension for Visual Studio (code writing and review, refactor, unit tests generation, documentation)</Description>
-        <License>LICENSE.txt</License>
-        <Icon>entwine-logo.ico</Icon>
-        <PreviewImage>Resources\vs-entwine-suggestion.png</PreviewImage>
-        <Tags>AI, artificial intelligence, assistant, coding, generator, llama, LLM, RAG</Tags>
-    </Metadata>
-    <Installation>
-        <InstallationTarget Id="Microsoft.VisualStudio.Community" Version="[17.0, 18.0)">
-            <ProductArchitecture>amd64</ProductArchitecture>
-        </InstallationTarget>
-    </Installation>
-    <Dependencies>
-        <Dependency Id="Microsoft.Framework.NDP" DisplayName="Microsoft .NET Framework" d:Source="Manual" Version="[4.5,)" />
-    </Dependencies>
-    <Prerequisites>
-        <Prerequisite Id="Microsoft.VisualStudio.Component.CoreEditor" Version="[17.0,18.0)" DisplayName="Visual Studio core editor" />
-    </Prerequisites>
-    <Assets>
-        <Asset Type="Microsoft.VisualStudio.VsPackage" d:Source="Project" d:ProjectName="%CurrentProject%" Path="|%CurrentProject%;PkgdefProjectOutputGroup|" />
-    </Assets>
+  <Metadata>
+    <Identity Id="EntwineLlm.77601dfd-d66f-405b-b3c4-c2e05497a5ca" Version="1.10" Language="en-US" Publisher="Emiliano Musso" />
+    <DisplayName>EntwineLlm</DisplayName>
+    <Description xml:space="preserve">LLM coding assistant extension for Visual Studio (code writing and review, refactor, unit tests generation, documentation)</Description>
+    <License>LICENSE.txt</License>
+    <Icon>entwine-logo.ico</Icon>
+    <PreviewImage>Resources\vs-entwine-suggestion.png</PreviewImage>
+    <Tags>AI, artificial intelligence, assistant, coding, generator, llama, LLM, RAG</Tags>
+  </Metadata>
+  <Installation>
+    <InstallationTarget Id="Microsoft.VisualStudio.Community" Version="[17.0, 18.0)">
+      <ProductArchitecture>amd64</ProductArchitecture>
+    </InstallationTarget>
+  </Installation>
+  <Dependencies>
+    <Dependency Id="Microsoft.Framework.NDP" DisplayName="Microsoft .NET Framework" d:Source="Manual" Version="[4.5,)" />
+  </Dependencies>
+  <Prerequisites>
+    <Prerequisite Id="Microsoft.VisualStudio.Component.CoreEditor" Version="[17.0,18.0)" DisplayName="Visual Studio core editor" />
+  </Prerequisites>
+  <Assets>
+    <Asset Type="Microsoft.VisualStudio.VsPackage" d:Source="Project" d:ProjectName="%CurrentProject%" Path="|%CurrentProject%;PkgdefProjectOutputGroup|" />
+  </Assets>
 </PackageManifest>
